refactor more buttons
