+++ baby-gru/src/components/button/MoorhenJedFlipTrueButton.tsx
@@ -1,0 +1,42 @@
+import { moorhen } from "../../types/moorhen";
+import { MoorhenContextButtonBase } from "./MoorhenContextButtonBase";
+import { MoorhenEditButtonBase } from "./MoorhenEditButtonBase";
+
+export const MoorhenJedFlipTrueButton = (props: moorhen.EditButtonProps | moorhen.ContextButtonProps) => {
+
+    const getCootCommandInput = (selectedMolecule: moorhen.Molecule, chosenAtom: moorhen.ResidueSpec, localParameters?: string): moorhen.cootCommandKwargs => {
+        return {
+            message: 'coot_command',
+            returnType: "status",
+            command: 'jed_flip',
+            commandArgs: [selectedMolecule.molNo, `//${chosenAtom.chain_id}/${chosenAtom.res_no}/${chosenAtom.atom_name}${chosenAtom.alt_conf === "" ? "" : ":" + chosenAtom.alt_conf}`, true],
+            changesMolecules: [selectedMolecule.molNo]
+        }
+    }
+
+    if (props.mode === 'context') {
+
+        return <MoorhenContextButtonBase 
+                    icon={<img style={{padding:'0.1rem', width:'100%', height: '100%'}} className="baby-gru-button-icon" src={`${props.urlPrefix}/baby-gru/pixmaps/jed-flip-reverse.svg`} alt='jed-flip-reverse'/>}
+                    toolTipLabel="JED Flip: wag the dog"
+                    cootCommandInput={getCootCommandInput(props.selectedMolecule, props.chosenAtom)}
+                    {...props}
+                />     
+
+    } else {
+
+        return <MoorhenEditButtonBase
+                    toolTip="JED Flip: wag the dog"
+                    buttonIndex={props.buttonIndex}
+                    selectedButtonIndex={props.selectedButtonIndex}
+                    setSelectedButtonIndex={props.setSelectedButtonIndex}
+                    needsMapData={false}
+                    getCootCommandInput={getCootCommandInput}
+                    prompt="Click atom in residue to flip around that rotatable bond - wag the dog"
+                    icon={<img style={{ width: '100%', height: '100%' }} alt="jed-flip-reverse" className="baby-gru-button-icon" src={`${props.urlPrefix}/baby-gru/pixmaps/jed-flip-reverse.svg`} />}
+                    {...props}
+                />
+    
+    }
+
+}
