+++ baby-gru/src/components/button/MoorhenSimpleEditButton.js
@@ -160,124 +160,6 @@ MoorhenSimpleEditButton.defaultProps = {
     awaitAtomClick: true, onExit: null, awaitMoreAtomClicksRef: false
 }
 
-export const MoorhenMutateButton = (props) => {
-
-    const autoFitRotamer = useCallback(async (molecule, chosenAtom) => {
-        const formattedArgs = [
-            molecule.molNo,
-            chosenAtom.chain_id,
-            chosenAtom.res_no,
-            chosenAtom.ins_code,
-            chosenAtom.alt_conf,
-            props.activeMap.molNo
-        ]
-        await props.commandCentre.current.cootCommand({
-            returnType: "status",
-            command: "auto_fit_rotamer",
-            commandArgs: formattedArgs,
-            changesMolecules: [molecule.molNo]
-        }, true)
-    }, [props.activeMap, props.commandCentre])
-
-    const [panelParameters, setPanelParameters] = useState("ALA")
-
-    const MoorhenMutatePanel = (props) => {
-        const toTypes = ['ALA', 'CYS', 'ASP', 'GLU', 'PHE', 'GLY', 'HIS', 'ILE',
-            'LYS', 'LEU', 'MET', 'ASN', 'PRO', 'GLN', 'ARG', 'SER', 'THR', 'VAL', 'TRP', 'TYR']
-        return <Container>
-            <Row>Please identify residue to mutate</Row>
-            <Row>
-                <FormGroup>
-                    <FormLabel>To residue of type</FormLabel>
-                    <FormSelect defaultValue={props.panelParameters}
-                        onChange={(e) => {
-                            props.setPanelParameters(e.target.value)
-                        }}>
-                        {toTypes.map(optionName => {
-                            return <option key={optionName} value={optionName}>{`${optionName} (${residueCodesThreeToOne[optionName]})`}</option>
-                        })}
-                    </FormSelect>
-                </FormGroup>
-            </Row>
-        </Container>
-    }
-
-    const mutateFormatArgs = (molecule, chosenAtom, pp) => {
-        return [molecule.molNo, `//${chosenAtom.chain_id}/${chosenAtom.res_no}`, pp]
-    }
-
-    return <MoorhenSimpleEditButton {...props}
-        id='mutate-residue-edit-button'
-        toolTip="Simple Mutate"
-        buttonIndex={props.buttonIndex}
-        selectedButtonIndex={props.selectedButtonIndex}
-        setSelectedButtonIndex={props.setSelectedButtonIndex}
-        needsMapData={true}
-        onCompleted={autoFitRotamer}
-        cootCommand="mutate"
-        panelParameters={panelParameters}
-        prompt={<MoorhenMutatePanel
-            setPanelParameters={setPanelParameters}
-            panelParameters={panelParameters} />}
-        icon={<img style={{ width: '100%', height: '100%' }} className="baby-gru-button-icon" src={`${props.urlPrefix}/baby-gru/pixmaps/mutate.svg`} alt='Mutate' />}
-        formatArgs={(m, c, p) => mutateFormatArgs(m, c, p)} />
-}
-
-export const MoorhenEigenFlipLigandButton = (props) => {
-    const [toolTip, setToolTip] = useState("Eigen Flip: flip the ligand around its eigenvectors")
-
-    useEffect(() => {
-        if (props.shortCuts) {
-            const shortCut = JSON.parse(props.shortCuts).eigen_flip
-            setToolTip(`Eigen Flip: flip the ligand around its eigenvectors ${getTooltipShortcutLabel(shortCut)}`)
-        }
-    }, [props.shortCuts])
-
-    return <MoorhenSimpleEditButton {...props}
-        id='eigen-flip-edit-button'
-        toolTip={toolTip}
-        buttonIndex={props.buttonIndex}
-        selectedButtonIndex={props.selectedButtonIndex}
-        setSelectedButtonIndex={props.setSelectedButtonIndex}
-        needsMapData={false}
-        cootCommand="eigen_flip_ligand"
-        prompt="Click atom in residue to eigen flip it"
-        icon={<img style={{ width: '100%', height: '100%' }} className="baby-gru-button-icon" src={`${props.urlPrefix}/baby-gru/pixmaps/spin-view.svg`} alt='Eigen flip' />}
-        formatArgs={(molecule, chosenAtom) => {
-            return [molecule.molNo, `//${chosenAtom.chain_id}/${chosenAtom.res_no}`]
-        }} />
-}
-
-export const MoorhenJedFlipFalseButton = (props) => {
-    return <MoorhenSimpleEditButton {...props}
-        toolTip="JED Flip: wag the tail"
-        buttonIndex={props.buttonIndex}
-        selectedButtonIndex={props.selectedButtonIndex}
-        setSelectedButtonIndex={props.setSelectedButtonIndex}
-        needsMapData={false}
-        cootCommand="jed_flip"
-        prompt="Click atom in residue to flip around that rotatable bond - wag the tail"
-        icon={<img style={{ width: '100%', height: '100%' }} className="baby-gru-button-icon" src={`${props.urlPrefix}/baby-gru/pixmaps/edit-chi.svg`} alt='jed-flip' />}
-        formatArgs={(molecule, chosenAtom) => {
-            return [molecule.molNo, `//${chosenAtom.chain_id}/${chosenAtom.res_no}/${chosenAtom.atom_name}${chosenAtom.alt_conf === "" ? "" : ":" + chosenAtom.alt_conf}`, false]
-        }} />
-}
-
-export const MoorhenJedFlipTrueButton = (props) => {
-    return <MoorhenSimpleEditButton {...props}
-        toolTip="JED Flip: wag the dog"
-        buttonIndex={props.buttonIndex}
-        selectedButtonIndex={props.selectedButtonIndex}
-        setSelectedButtonIndex={props.setSelectedButtonIndex}
-        needsMapData={false}
-        cootCommand="jed_flip"
-        prompt="Click atom in residue to flip around that rotatable bond - wag the dog"
-        icon={<img style={{ width: '100%', height: '100%' }} alt="jed-flip-reverse" className="baby-gru-button-icon" src={`${props.urlPrefix}/baby-gru/pixmaps/jed-flip-reverse.svg`} />}
-        formatArgs={(molecule, chosenAtom) => {
-            return [molecule.molNo, `//${chosenAtom.chain_id}/${chosenAtom.res_no}/${chosenAtom.atom_name}${chosenAtom.alt_conf === "" ? "" : ":" + chosenAtom.alt_conf}`, true]
-        }} />
-}
-
 export const MoorhenRotamerChangeButton = (props) => {
     const [showAccept, setShowAccept] = useState(false)
     const theButton = useRef(null)
